Re organized code builder heirarchy
