+++ hat/examples/squares/src/main/java/squares/Main.java
@@ -53,6 +53,7 @@ public class Main {
                 kc -> squareKernel(kc, s32Array)
         );
     }
+    
 
     public static void main(String[] args) {
         var lookup = java.lang.invoke.MethodHandles.lookup();
